- use unique_ptr instead of raw pointers - fix memory leak on ui modals - check if texture is valid before setting up dimensions - ensure texture is valid before removing - fix music leak - clean level objects if we cant determine level end - prevent loading invalid level if we fail loading level objects (could happen on 2 attempt) - invalid location on level save - bunch of missing casts - removed dead code - added extra flags for cmake (release opt, warns) q

+++ src/entity/world/spike.cpp
@@ -4,24 +4,24 @@
 
 #include <raylib.h>
 
-Spike::Spike(float ammount) : GameObject(ObjectType::SPIKE), m_ammount(ammount) {
+Spike::Spike(int ammount) : GameObject(ObjectType::SPIKE), m_ammount(ammount) {
     game.add_game_object(this);
     rb->is_static = true;
 
-    dimensions.x = ammount * 64.0f;
+    dimensions.x = static_cast<float>(ammount) * 64.0f;
     dimensions.y = 64.0f;
 }
 
 void Spike::render() {
     if (!visible) return;
 
-    float width = dimensions.x / m_ammount;
+    float width = dimensions.x / static_cast<float>(m_ammount);
 
     // Debug
     // DrawRectangleV(position, dimensions, RED);
 
     for (int i = 0; i < m_ammount; i++) {
-        float spacing = i == 0 ? 0 : i * width;
+        float spacing = i == 0 ? 0.0f : static_cast<float>(i) * width;
 
         DrawTriangle({position.x + spacing + width / 2, position.y}, {position.x + spacing, position.y + dimensions.y},
                      {position.x + spacing + width, position.y + dimensions.y}, {255, 0, 0, 255});
